Adds proper handling of Prismatic orientation.

+++ src/Vector.java
@@ -1,9 +1,9 @@
 import java.io.*;
 
 public class Vector implements Serializable {
-    public int m_x, m_y;
+    public double m_x, m_y;
 
-    public Vector(int x, int y) {
+    public Vector(double x, double y) {
         m_x = x;
         m_y = y;
     }
